Feat: 관리자 기수 등록, 검색, 수정, 삭제 API 구현

+++ api/generation/service.js
@@ -1,7 +1,7 @@
 const Generation = require('./model');
-// const { CustomError } = require('../../middlewares');
-// const { code, message } = require('../../utils');
 const { Op } = require('sequelize');
+const { getPagination, getPagingData } = require('../../utils');
+const { BadRequestClass, NotFoundClass } = require('../../middlewares');
 
 module.exports = {
   async addGeneration(generationInfo) {
@@ -10,32 +10,54 @@ module.exports = {
       where: { [Op.and]: [{ name: name }, { phase: phase }] },
     });
     if (generation) {
-      // throw new CustomError(code.CONFLICT, message.ALREADY_GENERATION);
-      throw new Error('이미 존재하는 기수입니다. 다시 한 번 확인해 주세요.');
+      throw new BadRequestClass('이미 존재하는 기수입니다.');
     }
 
     const createdNewGeneration = await Generation.create(generationInfo);
     return createdNewGeneration;
   },
 
-  async getGeneraions(keyword) {
-    const generations = await Generation.findAll({
-      where: {
-        [Op.or]: [
-          { name: { [Op.like]: `%${keyword}%` } },
-          { phase: { [Op.like]: `%${keyword}%` } },
+  async getGeneraions(page, size, keyword) {
+    const { limit, offset } = getPagination(page, size);
+    let generations;
+
+    if (!keyword || keyword === ':keyword') {
+      generations = await Generation.findAndCountAll({
+        order: [
+          ['name', 'ASC'],
+          ['phase', 'ASC'],
         ],
-      },
-    });
+        offset,
+        limit,
+      });
+    } else {
+      generations = await Generation.findAndCountAll({
+        where: {
+          [Op.or]: [
+            { name: { [Op.like]: `%${keyword}%` } },
+            { phase: { [Op.like]: `%${keyword}%` } },
+          ],
+        },
+        order: [
+          ['name', 'ASC'],
+          ['phase', 'ASC'],
+        ],
+        offset,
+        limit,
+      });
+    }
+
+    generations = getPagingData(generations, page, limit);
+
     return generations;
   },
 
   async setGeneration(id, toUpdate) {
     const { name, phase } = toUpdate;
+
     const generation = await Generation.findOne({ where: { id: id } });
     if (!generation) {
-      // throw new CustomError(code.NOT_FOUND, message.NO_GENERATION);
-      throw new Error('찾으시는 기수가 없습니다. 다시 한 번 확인해 주세요.');
+      throw new NotFoundClass('존재하지 않는 기수입니다.');
     }
 
     const updateCount = Generation.update(
@@ -48,21 +70,20 @@ module.exports = {
       }
     );
     if (updateCount < 1) {
-      // throw new CustomError(code.NOT_FOUND, message.GENERATION_UPDATE_FAIL);
-      throw new Error(
-        '기수 수정 처리가 실패했습니다. 다시 한 번 확인해 주세요.'
-      );
+      throw new BadRequestClass('기수 수정 처리에 실패했습니다.');
     }
     return updateCount;
   },
 
   async deleteGeneration(id) {
+    const generation = await Generation.findOne({ where: { id: id } });
+    if (!generation) {
+      throw new NotFoundClass('존재하지 않는 기수입니다.');
+    }
+
     const deleteCount = await Generation.destroy({ where: { id: id } });
     if (deleteCount < 1) {
-      // throw new CustomError(code.NOT_FOUND, message.GENERATION_DELETE_FAIL);
-      throw new Error(
-        '기수 삭제 처리가 실패했습니다. 다시 한 번 확인해 주세요.'
-      );
+      throw new BadRequestClass('기수 삭제 처리에 실패했습니다.');
     }
     return deleteCount;
   },
